+++ src/main/java/io/hhplus/tdd/controller/ApiControllerAdvice.java
@@ -1,5 +1,6 @@
-package io.hhplus.tdd;
+package io.hhplus.tdd.controller;
 
+import io.hhplus.tdd.controller.dto.ErrorResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
