+++ src/main/java/io/hhplus/tdd/controller/dto/ErrorResponse.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd;
+package io.hhplus.tdd.controller.dto;
 
 public record ErrorResponse(
         String code,
