+++ src/main/java/io/hhplus/tdd/domain/PointHistory.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd.point;
+package io.hhplus.tdd.domain;
 
 public record PointHistory(
         long id,
