+++ src/main/java/io/hhplus/tdd/domain/TransactionType.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd.point;
+package io.hhplus.tdd.domain;
 
 
 /**
